reorganize the licenses module a bit
