+++ pyproject.toml
@@ -60,7 +60,7 @@ ignore_missing_imports = true
 [tool.ruff]
 src = ["src"]
 extend-exclude = [
-    "src/packaging/licenses/spdx.py"
+    "src/packaging/licenses/_spdx.py"
 ]
 
 [tool.ruff.lint]
