+++ src/packaging/licenses/_spdx.py
